Allow packaging as a gem
