+++ Gemfile
@@ -1,3 +1,4 @@
-gem 'prometheus'
-gem 'prometheus-client'
-gem 'rack'
+source 'https://rubygems.org'
+
+# Specify your gem's dependencies in twemproxy_exporter.gemspec
+gemspec
