+++ lib/twemproxy_exporter/counter.rb
@@ -1,0 +1,19 @@
+module TwemproxyExporter
+  class Counter
+    def initialize(registry, name, desc)
+      @counter = Prometheus::Client::Counter.new(name, desc)
+      registry.register(@counter)
+      @last = 0
+    end
+
+    def count(value, labels = {})
+      if value >= @last
+        @counter.increment(labels, value - @last)
+      else
+        @counter.increment(labels, value)
+      end
+
+      @last = value
+    end
+  end
+end
