+++ lib/twemproxy_exporter/gauge.rb
@@ -1,0 +1,12 @@
+module TwemproxyExporter
+  class Gauge
+    def initialize(registry, name, desc)
+      @gauge = Prometheus::Client::Gauge.new(name, desc)
+      registry.register(@gauge)
+    end
+
+    def count(value, labels = {})
+      @gauge.set(labels, value)
+    end
+  end
+end
