+++ lib/twemproxy_exporter/version.rb
@@ -1,0 +1,3 @@
+module TwemproxyExporter
+  VERSION = "0.1.0"
+end
