+++ spec/twemproxy_exporter_spec.rb
@@ -1,0 +1,11 @@
+require "spec_helper"
+
+describe TwemproxyExporter do
+  it "has a version number" do
+    expect(TwemproxyExporter::VERSION).not_to be nil
+  end
+
+  it "does something useful" do
+    expect(false).to eq(true)
+  end
+end
